html/browsers/history/the-history-interface/history_pushstate_nooptionalparam.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>history_pushState_NoOptionalParam</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    try {
      window.history.pushState(1, document.title);

      var state;
      state = window.history.state;
      test(function() {
        assert_equals(state, 1, "pushState method is failed.");
      });
    } catch (e) {
      test(function() {
        assert_unreached("call of pushState method is failed.");
      });
    }

    </script>
  </body>
</html>
